+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Use OpenTelemetry semantic conventions for log correlation
 - Remove spans for Confluent.Kafka Consume calls that didn't receive a message
+- Add configuration setting, `SIGNALFX_OUTBOUND_HTTP_EXCLUDED_HOSTS`, that prevents the creation of outbound HTTP spans for certain hosts
 
 [Commits](https://github.com/signalfx/signalfx-dotnet-tracing/v0.1.12...HEAD)